fix: fix failed custom maps dialog confirmation on post story ds-33
